+++ Plugins/BDContextMenu/CHANGELOG.md
@@ -1,5 +1,11 @@
 # BDContextMenu Changelog
 
+### 0.0.2
+
+ - Add item-group
+ - Add submenus to everything
+ - Allow enable/disable of plugins, themes, core, and emotes!
+
 ### 0.0.1
 
  - Initial version